refactor: rename NestJS module and service files and update test directory structure.
