+++ packages/core/package.json
@@ -21,7 +21,7 @@
   "scripts": {
     "quick:test": "cross-env NODE_NO_WARNINGS=1 npx tsx ../../tests/setup.ts",
     "test": "c8 pnpm quick:test",
-    "lint": "eslint 'src/**/*.ts' --fix",
+    "lint": "eslint 'src/**/*.ts' 'tests/**/*.ts' --fix",
     "build": "tsc"
   },
   "dependencies": {
