+++ packages/nestjs/src/service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Inject, Logger } from "@nestjs/common";
+import { Injectable, Inject } from "@nestjs/common";
 import { createPermifyClient } from "@permify-toolkit/core";
 
 import { PERMIFY_MODULE_OPTIONS } from "./constant";
@@ -7,15 +7,11 @@ import type { PermifyModuleOptions } from "./interfaces";
 @Injectable()
 export class PermifyService {
   readonly client: ReturnType<typeof createPermifyClient>;
-  private readonly logger = new Logger(PermifyService.name);
 
   constructor(
     @Inject(PERMIFY_MODULE_OPTIONS)
     private readonly options: PermifyModuleOptions
   ) {
-    this.logger.log(
-      `Initializing Permify client with endpoint: ${options.client.endpoint}`
-    );
     this.client = createPermifyClient(options.client);
   }
